Create Interviewportal
This project is aimed at developing an online search Portal for the
Placement Details for job seekers. The system is an online application that can be accessed throughout the organization and inside as well with proper login provided. This system can be used as an Online Job Portal for job seekers. Job Seekers logging should be able to upload their information in the form of a CV. Visitors/Company representatives logging in may also access/search any information put up by Job aspirants